% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_entropy}
\alias{calc_entropy}
\title{Calculate Shannon's Entropy}
\usage{
calc_entropy(ra_mat, priors = NULL)
}
\arguments{
\item{ra_mat}{A matrix with rows equal to the number of rules and columns
equal to the number of actions}

\item{priors}{A vector of prior probabilities. Defaults to NULL.}
}
\value{
Returns a vector of entropies for each possible action with the
following attributes:
\enumerate{
  \item ra_mat
  \item priors
  \item pr_aj_rk
  \item pr_rk_aj
}
}
\description{
This function is used to calculate Shannon's Entropy. The objective of the 
decision maker is to maximize the entropy of the observer.
}
