% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_payouts.R
\name{calc_payout_obs}
\alias{calc_payout_obs}
\title{Calculate expected payout to the observer}
\usage{
calc_payout_obs(pr_rk_aj, pay_obs)
}
\arguments{
\item{pr_rk_aj}{A matrix of posterior probabilities}

\item{pay_obs}{The pay to the observer for guessing correctly.}
}
\value{
A vector of expected pays for each possible guess
}
\description{
The function calculates the expected payout to the observer. The payout to
the observer depends on the posterior probabilities, i.e. the probability
of a rule conditional on observing an action, and the monetary payout for 
guessing correctly. The function is meant for internal use only. To calculate
the payout to the observer, use the wrapper function 
\code{\link{calculate_payouts}}.
}
