% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_pr_aj_rk}
\alias{calc_pr_aj_rk}
\title{Calculate Pr(a_j|r_k)}
\usage{
calc_pr_aj_rk(ra_mat)
}
\arguments{
\item{ra_mat}{A matrix with rows equal to the number of rules and columns
equal to the number of actions}
}
\value{
An r x a matrix of probabilities
}
\description{
The function calculates the probability of an action conditional on a given
rule. The function is for internal use only and cannot be called externally.
The resulting matrix will be supplied as an attribute to the entropy measure
see \code{\link{calc_entropy}}.
}
\references{
Equation X in Chorus et al. (xxxx)
}
