% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_pr_rk_aj}
\alias{calc_pr_rk_aj}
\title{Calculate Pr(r_k|a_j)}
\usage{
calc_pr_rk_aj(pr_aj_rk, priors)
}
\arguments{
\item{pr_aj_rk}{A matrix with the probabilities of actions conditional on a
given rule. Calculated using \code{\link{calc_pr_aj_rk}}}

\item{priors}{A vector of prior probabilities. Defaults to NULL.}
}
\value{
An r x a matrix of probabilities
}
\description{
The function calculates the probability of a rule conditional on observing
a given action. The function is for internal use only and cananot be called
externally. The resulting matrix will be supplied as an attribute to the 
entropy measure see \code{\link{calc_entropy}}.
}
\references{
Equation X in Chorus et al. (xxxx)
}
