% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calculate_pay_decision_maker}
\alias{calculate_pay_decision_maker}
\title{Function for calculaton the pay to the decision maker}
\usage{
calculate_pay_decision_maker(prob_guessing, pay_decision_maker)
}
\arguments{
\item{prob_guessing}{The probability that the observer will make a guess. This vector is passed from \code{\link{calculate_prob_guess}}.}

\item{pay_decision_maker}{The pay to the decision maker if the observer does not make a guess.}
}
\value{
A vector of expected pays for each possible guess
}
\description{
This function is used to calculate the expected pay to the decision maker
}
