% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calculate_pay_observer}
\alias{calculate_pay_observer}
\title{Function for calculaton the pay to the observer}
\usage{
calculate_pay_observer(prob_rule_action, pay_observer)
}
\arguments{
\item{prob_rule_action}{A matrix containing the conditional probability of a rule given an observed action.
This matrix is an output from \code{\link{calculate_entropy}}.}

\item{pay_observer}{The pay to the observer for guessing correctly. Usually passed through design_opt}
}
\value{
A vector of expected pays for each possible guess
}
\description{
This function is used to calculate the expected pay to the observer
}
