% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calculate_prob_guess}
\alias{calculate_prob_guess}
\title{Function for calculating the probability of guessing}
\usage{
calculate_prob_guess(expected_pay_observer, pay_observer_no_guess,
  design_opt)
}
\arguments{
\item{expected_pay_observer}{Vectro of expected pay calculated by \code{\link{calculate_pay_observer}}.}

\item{pay_observer_no_guess}{The payout to the observer if she refrains from guessing}

\item{design_opt}{A list of design options that govern the creation of the obfuscation games.}
}
\description{
This function is used to calculate the probability that the observer will make a guess 
as to what rule governs a decision maker's action.
}
