% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ra_mat.R
\name{construct_design}
\alias{construct_design}
\title{Function to create a rule-action matrix}
\usage{
construct_design(design_opt)
}
\arguments{
\item{design_opt}{List of design options}
}
\value{
A rules-action matrix
}
\description{
The function creates a rule-action matrix (i.e. an obfuscation design)
subject to a list of pre-programmed restrictions. These restrictions are in
place to ensure that no invalid designs are created. Some of these restrictions
can be changed by the user by appropriately specifying the 
\code{design_opt_input}. Each matrix is a design for one period of the 
the obfuscation game. This function is for internal use only. To create
an obfuscation design, the user should use \code{\link{generate_designs}}.
}
