% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{construct_ra_mat}
\alias{construct_ra_mat}
\title{Function to create a rule-action matrix}
\usage{
construct_ra_mat(design_opt)
}
\arguments{
\item{design_opt}{A list of design options}
}
\value{
A rules-action matrix
}
\description{
The function takes as an input a list of design options and given a set of 
pre-programmed restrictions outputs one rule-action matrix. This matrix
is the design for one period of the obfuscation game. The package comes with
several other functions that are useful to evaluate the ra-matrix. The games 
will need to be manually inspected prior to implementation.
}
\details{
The function is for internal use only and not exported
}
