% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_opt.R
\docType{data}
\name{design_opt}
\alias{design_opt}
\title{Elements of the design matrix}
\format{An object of class \code{list} of length 16.}
\usage{
design_opt
}
\description{
\describe{
  \item{rules}{Number of rules (i.e. rows)}
  \item{actions}{Number of actions (i.e. columns)}
  \item{considered_rule}{The considered rule, i.e. specifies the row}
  \item{min}{Minimum number of actions available for the considered rule}
  \item{max}{Maximum number of actions available for the considered rule}
  \item{min_fit}{Minimum number of rules fitting each permitted action
  conditional on the rule}
  \item{obligatory}{Number of rules with obligatory actions}
  \item{sd_entropy}{Specifies the standard deviation of the entropy values}
  \item{pay_obs}{The pay to the observer for guessing}
  \item{pay_obs_no_guess}{The pay to the observer for not guessing}
  \item{deterministic}{Is the observer guessing deterministic}
  \item{pay_dm}{The pay to the decision maker if the observer does not guess}
  \item{designs}{Number of designs to generate}
  \item{max_iter}{Maximum number of iterations before stopping search for
  designs}
  \item{print_all}{Print all information about the design}
  \item{seed}{A seed for the random number generator. Useful for
  replicability}
  
}
}
\keyword{datasets}
