% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_attach.R
\name{.onAttach}
\alias{.onAttach}
\title{Print package startup message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\value{
Nothing
}
\description{
The function is called when the package is loaded through library or require.
}
