% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_designs.R
\name{generate_designs}
\alias{generate_designs}
\title{Generate obfuscation designs}
\usage{
generate_designs(design_opt_input = list())
}
\arguments{
\item{design_opt_input}{A list of user supplied design options. The default
value is an empty list.}
}
\value{
A list of matrices with rules and actions
}
\description{
This is the main function to generate the obfuscation designs. It takes a 
list of design options as its only argument. See 
\code{\link{check_design_opt}} for default values.
}
\examples{
#   Example 1 - Using only default values
generate_designs()

#   Example 2 - Using some user supplied values
design_opt_input <- list(rules = 5,
                         actions = 5,
                         considered_rule = 5,
                         min = 2, 
                         max = 3,
                         min_fit = 2,
                         obligatory = 2,
                         sd_entropy = 0.15)


generate_designs(design_opt_input)

}
