% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{last}
\alias{last}
\title{Get the last element of a vector}
\usage{
last(x)
}
\arguments{
\item{x}{A vector}
}
\description{
\code{last} extracts the last element of a vector
}
\examples{
x <- 1:4
last(x)

x <- c("hello", "my", "name", "is", "buttons")
last(x)

}
