% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obfuscatoR-package.R
\docType{package}
\name{obfuscatoR-package}
\alias{obfuscatoR}
\alias{obfuscatoR-package}
\title{obfuscatoR: Obfuscation Game Designs}
\description{
When people make decisions, they may do so using a wide variety of decision rules. The package allows users to easily create obfuscation games to test the obfuscation hypothesis. It provides an easy to use interface and multiple options designed to vary the difficulty of the game and tailor it to the user's needs. For more detail: Chorus et al., 2021, Obfuscation maximization-based decision-making: Theory, methodology and first empirical evidence, Mathematical Social Sciences, 109, 28-44, \doi{10.1016/j.mathsocsci.2020.10.002}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://obfuscator.edsandorf.me}
  \item \url{https://github.com/edsandorf/obfuscatoR}
}

}
\author{
\strong{Maintainer}: Erlend Dancke Sandorf \email{erlend.dancke.sandorf@nmbu.no}

Authors:
\itemize{
  \item Caspar Chorus
  \item Sander van Cranenburgh
}

}
\keyword{internal}
