% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obfuscatoR.R
\docType{package}
\name{obfuscatoR}
\alias{obfuscatoR}
\alias{obfuscatoR-package}
\title{obfuscatoR: Designs and analysis of the obfuscation game}
\description{
This package creates designs for the obfuscation game. It has a simple 
interface where you specify the number of strategies and outcomes. Given a  
set of constraints, feasible designs are provided that can be readily  
implemented to replicate results or extend the game to test new types of  
behavior.
}
