% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print_object}
\alias{print_object}
\title{Function for printing vectors and matrices}
\usage{
print_object(X, row_wise = FALSE, digits = 2)
}
\arguments{
\item{X}{Input matrix}

\item{row_wise}{If TRUE populate something row_wise. Default FALSE}

\item{digits}{Number of digits for rounding}
}
\description{
Function for printing vectors and matrices
}
