% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print_payout}
\alias{print_payout}
\title{Print the payouts}
\usage{
print_payout(payout, digits = 3, print_all = FALSE)
}
\arguments{
\item{payout}{A list of calculated payouts}

\item{digits}{The number of digits to round to. Default 3.}

\item{print_all}{If TRUE will print the probabilities of guessing}
}
\description{
Print the payouts
}
