% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print_ra}
\alias{print_ra}
\title{Prints the rules-action matrix}
\usage{
print_ra(ra_mat, print_all = FALSE)
}
\arguments{
\item{ra_mat}{A matrix with rows equal to the number of rules and columns
equal to the number of actions}

\item{print_all}{If TRUE prints information on the number of iterations and 
and whether all design conditions were met. Default is FALSE}
}
\description{
Prints the rule and action matrix. Depending on options, additional text is 
provided with information on the considered rule and/or the design generation 
process.
}
\examples{
design_opt_input <- list(rules = 4, 
                         actions = 5)

design <- generate_designs(design_opt_input)

print_ra(design)
print_ra(design, TRUE)

}
