% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print_ra_mat}
\alias{print_ra_mat}
\title{Prints the ra_matrix}
\usage{
print_ra_mat(ra_mat, c_rule = NULL, all_info = FALSE)
}
\arguments{
\item{ra_mat}{A matrix with rows equal to the number of rules and columns
equal to the number of actions}

\item{c_rule}{Defaults to the rule specified in design_opt.}

\item{all_info}{If TRUE prints informatino on the number of iterations and 
and whether all design conditions have been met. Default is FALSE}
}
\description{
Prints the ra_matrix with the considered rule highlighted
}
