% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_design.R
\name{save_design}
\alias{save_design}
\title{Save obfuscation designs}
\usage{
save_design(x, x_name, path = getwd())
}
\arguments{
\item{x}{A design or list of designs}

\item{x_name}{A character string with the name of the file}

\item{path}{A string giving the path to where the designs are stored. The 
default is the current working directory}
}
\value{
The stored .csv file has a naming structure that has the number of
the design and the considered rule as part of the file name.
}
\description{
Writes a design to a .csv file. If supplied a list of designs, it writes
the designs to separate .csv files.
}
