% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_payout_dm}
\alias{calc_payout_dm}
\title{Calculate expected payout to the decision maker}
\usage{
calc_payout_dm(pr_guess, pay_dm)
}
\arguments{
\item{pr_guess}{A vector with the probability that the observer will make a 
guess.}

\item{pay_dm}{The pay to the decision maker if the observer does 
not make a guess.}
}
\value{
A vector of expected payouts for each possible guess made by the
 observer
}
\description{
The function calculates the expected payout to the decision maker.
}
