% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_payout_obs}
\alias{calc_payout_obs}
\title{Calculate expected payout to the observer}
\usage{
calc_payout_obs(pr_rk_aj, pay_obs)
}
\arguments{
\item{pr_rk_aj}{A matrix with the probabilities of a rule conditional on an
observed action}

\item{pay_obs}{The pay to the observer for guessing correctly.}
}
\value{
A vector of expected pays for each possible guess
}
\description{
The function calculates the expected payout to the observer.
}
