% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_entropy.R
\name{calc_pr_rk_aj}
\alias{calc_pr_rk_aj}
\title{Calculate Pr(r_k|a_j)}
\usage{
calc_pr_rk_aj(pr_aj_rk, priors)
}
\arguments{
\item{pr_aj_rk}{A matrix with the probabilities of actions conditional on a
given rule.}

\item{priors}{A vector of prior values. If the design is a list of matrices,
priors can be a matrix with rows equal to the length of the design and columns
equal to the number of rules.}
}
\value{
An r x a matrix of probabilities
}
\description{
The function calculates the probability of a rule conditional on observing
a given action and is part of calculating the entropy of an action. This 
probability is also referred to as the posterior probability. The function
is meant for internal use only.
}
