% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Function for calculating Shannon's Entropy}
\usage{
calculate_entropy(rule_action_matrix, priors = NULL)
}
\arguments{
\item{rule_action_matrix}{A matrix with rows equal to the number of rules and columns equal to the number of actions}

\item{priors}{A vector of priors. Defaults to NULL and uninformative priors 1/R is used.}
}
\description{
This function is used to calculate Shannon's Entropy. The objective of the decision maker is to maximize the entropy of the observer.
}
