% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Calculate the entropy}
\usage{
calculate_entropy(ra_mat, priors = NULL)
}
\arguments{
\item{ra_mat}{A matrix with rows equal to the number of rules and columns
equal to the number of actions or a list of such matrices.}

\item{priors}{A vector of prior values. If ra_mat is a list of matrices,
priors can be a matrix with rows equal to the length of ra_mat and columns
equal to the number of rules.}
}
\value{
A list of of vectors of entropies for each possible action with the
following attributes:
\enumerate{
  \item ra_mat
  \item priors
  \item pr_aj_rk
  \item pr_rk_aj
}
}
\description{
This function is a wrapper for calc_entropy.
}
