% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design_opt.R
\name{check_design_opt}
\alias{check_design_opt}
\title{Check design options}
\usage{
check_design_opt(design_opt_input)
}
\arguments{
\item{design_opt_input}{A list of user supplied design options.}
}
\value{
Returns a list of design options with the missing from input replaced
by default values
}
\description{
The function checks the list of design options specified by the user and 
sets sensible defaults where no option is specified. The function is meant
for internal use only and is not exported to be used by the users. All 
options can be overridden by the the user by appropriately specifying 
\code{design_opt_input}. 
 
Below is a list defining each of the options available to be specified in 
\code{design_opt_input}.
}
\details{
\describe{
  \item{rules}{Number of rules (i.e. rows)}
  \item{actions}{Number of actions (i.e. columns)}
  \item{min}{Minimum number of actions available for the considered rule}
  \item{max}{Maximum number of actions available for the considered rule}
  \item{min_fit}{Minimum number of rules fitting each permitted action
  conditional on the rule}
  \item{obligatory}{Number of rules with obligatory actions}
  \item{sd_entropy}{Specifies the standard deviation of the entropy values}
  \item{designs}{Number of designs to generate}
  \item{max_iter}{Maximum number of iterations before stopping search for
  designs}
  \item{seed}{A seed for the random number generator. Useful for
  replicability}
}
}
