% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ra_mat.R
\name{construct_ra_mat}
\alias{construct_ra_mat}
\title{Function to create a rule-action matrix}
\usage{
construct_ra_mat(design_opt)
}
\arguments{
\item{design_opt}{List of design options}
}
\value{
A rules-action matrix
}
\description{
The function takes as an input a list of design options and given a set of 
pre-programmed restrictions outputs one rule-action matrix. This matrix
is the design for one period of the obfuscation game. The package comes with
several other functions that are useful to evaluate the ra-matrix. The games 
will need to be manually inspected prior to implementation.
}
\details{
The function is for internal use only and not exported
}
\examples{
design_opt_input <- list(rules = 5,
                         actions = 5,
                         considered_rule = 5,
                         min = 2, 
                         max = 3,
                         min_fit = 2,
                         obligatory = 2,
                         sd_entropy = 0.15)

#   Check design_opt_input
design_opt <- check_design_opt(design_opt_input)

construct_ra_mat(design_opt)

}
