% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_designs.R
\name{generate_designs}
\alias{generate_designs}
\title{Generate obfuscation designs}
\usage{
generate_designs(design_opt_input = list())
}
\arguments{
\item{design_opt_input}{A list of user supplied design options. Must at a 
minimum specify the number of rules, actions and the considered rule.}
}
\value{
A list of matrices with rules and actions
}
\description{
This is the main function to generate the obfuscation designs. It takes a 
list of design options as its only argument. See 
\code{\link{check_design_opt}} for default values.
}
\details{
At the very minimum, the number of rules, actions and the considered rule
has to be supplied
}
\examples{
design_opt_input <- list(rules = 4,
                         actions = 5)

generate_designs(design_opt_input)

}
