% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obfuscatoR.R
\docType{package}
\name{obfuscatoR}
\alias{obfuscatoR}
\alias{obfuscatoR-package}
\title{obfuscatoR: Designs and analysis of the obfuscation game}
\description{
When people make decisions, they may do so using a wide variety of decision 
rules. The package allows users to easily create obfuscation games to test 
the obfuscation hypothesis. It provides an easy to use interface and multiple 
options designed to vary the difficulty of the game and tailor it to the 
user's needs.
}
