% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print_object}
\alias{print_object}
\title{Function for printing vectors and matrices}
\usage{
print_object(X, row_wise = FALSE, digits = 2)
}
\arguments{
\item{X}{A vector or matrix to be printed}
}
\description{
Function for printing vectors and matrices
}
